feat(sdks): regenerate ABI bindings, reroute allocation, kill host statics
- all sdks/*/abi regenerated for the 144/80/24-byte ABI
- cpp host: on_reload functor flows through on_reload_user_data
  trampoline (static callback storage removed); warning-free build
- cpp guest: operator new/delete + alloc_string via stored
  HostInterface alloc/free; alloc-tracking test mocks a HostInterface
- rust guest: alloc_string through host vtable; cross-crate allocator
  re-exports removed (each dylib no longer links its own allocator)
- csharp: PinnedUtf8/StaticMessage string lifetimes, AbiError u32 codes
- js: by-value struct FFI for Array/load_bundle, reload callback takes
  user-data pointer; lua: single canonical runtime cdef
- python READMEs document the real generated-contracts + set_impl
  registration pattern

+++ sdks/cpp/loaders/dotnet/polyplug_loaders_dotnet.hpp
@@ -28,7 +28,7 @@ inline void register_dotnet(Runtime& rt, std::string_view min_framework = "10.0"
     static const char runtime_name[] = "dotnet";
     StringView name{reinterpret_cast<const uint8_t*>(runtime_name), sizeof(runtime_name) - 1};
     AbiError err = host->register_loader(host, name, loader);
-    if (err.code != AbiErrorCode::Ok) {
+    if (err.code != static_cast<uint32_t>(AbiErrorCode::Ok)) {
         throw std::runtime_error("polyplug: dotnet loader register failed: " + rt.get_last_error());
     }
 }
