+++ sdks/cpp/loaders/js/polyplug_loaders_js.hpp
@@ -23,7 +23,7 @@ inline void register_js(Runtime& rt) {
     static const char runtime_name[] = "js-quickjs";
     StringView name{reinterpret_cast<const uint8_t*>(runtime_name), sizeof(runtime_name) - 1};
     AbiError err = host->register_loader(host, name, loader);
-    if (err.code != AbiErrorCode::Ok) {
+    if (err.code != static_cast<uint32_t>(AbiErrorCode::Ok)) {
         throw std::runtime_error("polyplug: js loader register failed: " + rt.get_last_error());
     }
 }
